Version 3.12, add NLP (#42)
## Description

This PR pins the version to Python 3.12, which is required by one of the dependencies of `sentence-transformers`. It also adds the NLP and ML Packages to `pyproject.toml` for spinning up locally and checking memory footprint. During investigation of versioning and dependencies, a number of numpy compatibility issues were discovered, so those are also handled via explicit pins and dockerfile changes.