Add function to retrieve manifest data (#44)
## Description

Based on our current understanding of how TTC will integrate with APHL's
setup, we think that the APHL bridge will send an event to the TTC
lambda function when an eICR needs to be standardized. This PR adds the
functionality for the TTC lambda function to accept an EventBridge event
and return the files listed in the record(s).

## Related Issues
Closes #36

## Additional Notes
I used `moto` to mock the AWS functionality for our tests.