Update CODEOWNERS + PR template (#45)
## Description
I noticed that the Codeowners are not automatically being assigned as
reviewers. I think this minor fix to the file will do it. Also slightly
modified the PR template to encourage linking the issues in Zenhub so
that they close automatically when the PR is merged.

## Related Issues

n/a

## Additional Notes
I added a new ruleset to the codebase as well to automatically assign
CodeOwners as reviewers