Create script to grab valuesets and store them in CSV files for TTC Model (#41)
## Description

Created a script to load the various valuesets required for the eICR
fields that will be processed in the Text-to-Code solution for the MVP.
This script calls the various APIs necessary to gather the code &
text/terms for the specified valuesets and stores them in respective csv
files in our project.

### ValueSets & Fields:
* Lab Orders - Loinc - READY (for later use)
* Lab Observations - Loinc - READY (for later use)
* Lab Names (Both Ordering and Resulting) - Loinc - READY
* Lab Result Values - Snomed (Lab Results) - READY
* Lab Result Interpretations - HL7 Lab Interpretations - READY

## Related Issues


[GitHubIssue-30](https://app.zenhub.com/workspaces/dibbs-text-to-code-68756bf8ea06bb00319391c0/issues/gh/cdcgov/dibbs-text-to-code/30)

## Additional Notes

The LOINC and SNOMED (UMLS) APIs require Auth. - I decided to use
environment variables to pull these values. We can update this later to
use GitHub secrets if we would like. In the interim, if anyone on the
team needs a username/pwd or API Key, please reach out to me and I can
provide them.

I am including the resulting csv files in this PR as well so they can be
leverage in our current work efforts. If we want to update these files,
we can do so at that time by running the script again.

## Checklist

Please review and complete the following checklist before submitting
your pull request:

- [X] I have ensured that the pull request is of a manageable size,
allowing it to be reviewed within a single session.
- [X] I have reviewed my changes to ensure they are clear, concise, and
well-documented.
- [X] I have updated the documentation, if applicable.
- [X] I have added or updated test cases to cover my changes, if
applicable.
- [X] I have minimized the number of reviewers to include only those
essential for the review.

## Checklist for Reviewers

Please review and complete the following checklist during the review
process:

- [ ] The code follows best practices and conventions.
- [ ] The changes implement the desired functionality or fix the
reported issue.
- [ ] The tests cover the new changes and pass successfully.
- [ ] Any potential edge cases or error scenarios have been considered.

---------

Co-authored-by: Eric Buckley <eric.buckley@gmail.com>