Add word-level scrambler function (#72)
## Description
Adds a function to scramble text at the word level, ensuring that only
unique indices are swapped.

Also removed Eric as a code owner so that he won't be tagged in PRs and
updated the handler event to be an SQS Event based on our conversation
with Geo earlier this week. I moved `augmentation.py` into the
`dibbs_text_to_code` directory instead of nesting it under
`data_curation_and_augmentation` because it made the imports nicer but
let me know if you feel strongly about the nesting.

## Related Issues
Closes #66
Fixes #69 

## Additional Notes
Also removed Eric as a code owner so that he won't be tagged in PRs and
updated the handler event to be an SQS Event based on our conversation
with Geo earlier this week.