Generate positive pairs (#87)
## Description

This PR creates a function that generates positive pairs of given LOINC
data and writes them to a file for use in model training.

## Related Issues

Closes #61 

## Additional Notes:

While I couldn't test it on the LOINC files that we haven't made yet, I
did some testing on some dummy files and directories on my local (since
it's a function that produces file output and will be run one-off I
didn't see much point in unit tests beyond making sure it works).

That means this PR is actually **not blocked** by other dependencies,
we'll just need to wire up the file paths once that's done.