Write generate_augmented_examples function (#84)
## Description

This PR introduces an "orchestrator" function that takes as input a
standard LOINC code string and generates a list of augmented training
examples by probabilistically using other augmentation and scrambling
functions we have thus far developed.

## Related Issues

Closes #63 

## Additional Notes

Merging this PR is blocked until we get the enhancement function done
(so that I can refactor to add it, as well as test it), but all other
functionality and logic is as it should be. Probably most relevant for
the time being are the schemas for augmentation configurations and the
initial default config objects I'm proposing.

---------

Co-authored-by: m-goggins <marcelle.goggins@gmail.com>