Create LOINC Data Files and Validation Positive Pairs (#102)
## Description

This PR does several main things:

* Refactors the augmentation code to address a number of bugs; I won't
list them all (many of them are in the code comments as new
explanations), but we had problems in our character deletion and
throughout the enhancement code. That's now fixed and operating better
than before (notably because we now jointly model both single-token
enhancements and substrings).
* Updates unit tests accordingly
* Generates a set of LOINC data files which can be used for positive
pair creation
* Generates a new validation set of positive-pair data which can be run
through our performance notebook
* Refactors the performance notebook into performance and embedding so
that the functionality of each is more modular and independent

## Related Issues

Closes #104 

## Additional Notes

Apologies for how much got bundled into this PR! I didn't intend to
creep it up with so much new code, but between the bugs/refactor and a
lot of quality of life stuff...well, here we are 😅