Add Azure ML Environment file (#123)
## Description
This PR adds a .yml environment file that can be attached to Azure ML
compute instances to hopefully improve package stability. I added a
`README` to the `azure_scripts` directory that includes details on how
to attach the environment to a compute instance. You only need to attach
the environment one time and then it should be available for all future
use.

## Related Issues

Closes #112

## Additional Notes
I looked into using a Dockerfile instead of conda, but Azure ML Studio
doesn't allow you to attach anything but a conda file to a compute
instance (at least as far as I can tell). It seems that you can attach a
Dockerfile environment to an MLClient `job` but since we're not using
jobs at this point in time, I went with the conda set up.

<--------------------- REMOVE THE LINES BELOW BEFORE MERGING
--------------------->

## Checklist

Please review and complete the following checklist before submitting
your pull request:

- [x] I have ensured that the pull request is of a manageable size,
allowing it to be reviewed within a single session.
- [x] I have reviewed my changes to ensure they are clear, concise, and
well-documented.
- [x] I have updated the documentation, if applicable.
- [x] I have added or updated test cases to cover my changes, if
applicable.
- [x] I have minimized the number of reviewers to include only those
essential for the review.

## Checklist for Reviewers

Please review and complete the following checklist during the review
process:

- [ ] The code follows best practices and conventions.
- [ ] The changes implement the desired functionality or fix the
reported issue.
- [ ] The tests cover the new changes and pass successfully.
- [ ] Any potential edge cases or error scenarios have been considered.