Add AWS-Opensearch to docker file - Using the new Project Structure (#130)
## Description

Added the ability to spin up an opensearch container, via our docker
compose file. Basic settings are included and some of the
parameters/configurations may need to be tweaked as we start to actually
use it. I have added a healthcheck as well.

**NOTE: You will need to set an environment variable `OPENSEARCH_PWD`
for the admin password for opensearch**
## Related Issues

Closes #108 

## Additional Notes

<--------------------- REMOVE THE LINES BELOW BEFORE MERGING
--------------------->

## Checklist

Please review and complete the following checklist before submitting
your pull request:

- [x] I have ensured that the pull request is of a manageable size,
allowing it to be reviewed within a single session.
- [x] I have reviewed my changes to ensure they are clear, concise, and
well-documented.
- [x] I have updated the documentation, if applicable.
- [x] I have added or updated test cases to cover my changes, if
applicable.
- [x] I have minimized the number of reviewers to include only those
essential for the review.

## Checklist for Reviewers

Please review and complete the following checklist during the review
process:

- [ ] The code follows best practices and conventions.
- [ ] The changes implement the desired functionality or fix the
reported issue.
- [ ] The tests cover the new changes and pass successfully.
- [ ] Any potential edge cases or error scenarios have been considered.