Add embedding notebook (#125)
## Description

This PR adds the embedding notebook from Azure into our repo. Not that
anyone would try to run it locally, but just as a note it's unlikely to
work outside of the cloud due to some `assert` statements I built in
around `torch.cuda` (unless you have a really beefy GPU).

## Related Issues

Closes #114 
Closes #131 

## Additional Notes

Embedding with even the small crappy GPU that's super cheap is 20-30x
faster than embedding on my local so we are _cooking_