Add LOINC Lab Type to Extracts (#143)
## Description

Added new column to the LOINC Lab extract files and processes. This new
field/column is 'lab_type' which stores the orderobs value from LOINC to
indicate what type of lab it is. The various categories are
'Observation' (aka. Result/Resulting), 'Order' (aka. Ordering), and
'Both'. This is necessary so that only the LOINC codes that are 'Order'
and 'Both' are used in comparison and returned, by the TTC application,
as a possible LOINC Code for Lab Name (Ordering) section of the eICR.
And to ensure that only the LOINC codes that are 'Observation' and
'Both' are used in comparison and returned, by the TTC application, as a
possible LOINC Code for Lab Name (Resulting) section of the eICR.

I have also updated the documentation for data curation to match the new
structures introduced into our project repo.

## Related Issues

Closes #144 

## Additional Notes

## Checklist for Reviewers

Please review and complete the following checklist during the review
process:

- [ ] The code follows best practices and conventions.
- [ ] The changes implement the desired functionality or fix the
reported issue.
- [ ] The tests cover the new changes and pass successfully.
- [ ] Any potential edge cases or error scenarios have been considered.