Creating jsonl embeddings (#153)
## Description

Adds a script to run after the embedding scripts to parse out 1000-line
long `.jsonl` files for each model.

So far only ran for qwen. A few of the other pickles I tried to load
were corrupted when I tried to load them, not sure if others are
encountering the same issue for the ellen models.

Sample jsonls for qwen:
https://drive.google.com/drive/u/0/folders/1GZIPJ8uok3A9PQcdOeMhy1Nz-onaiYzF


## Related Issues

Closes #148