Fix embeddings with empty 'code' element - add loinc types to all embeddings - split embeddings (#170)
## Description

While adding the LOINC types to the embedding files an issue was found
where some of the 'codes' (actually the LOINC text descriptions for
short_name and display_name) were empty ("") in the embedding files. We
also weren't comparing for short name and display name. I updated the
LOINC Type notebook to recreate the embedding files without any 'code'
entries that were null/empty. I also ensured that every record had a
loinc type associated with it.

I have include the script and resulting file that provided the data
about how many loinc codes had either an empty short_name, display_name,
or both. It also calculates how many embeddings were created with null
'codes'. We may want to move the script and the resulting file to a
different location??

The notebook has successfully completed and the results can be found
[here](https://dibbsttc9073465248.blob.core.windows.net/azureml-blobstore-d49d62ee-fc90-4e2a-add1-39879de70ed8/embeddings/refined/).

I also worked with Rob to get his 'split' script into a notebook that we
could run against the newly updated embedding files. I have included
that notebook in this PR as well.

This notebook has also successfully completed and the results can be
found
[here](https://dibbsttc9073465248.blob.core.windows.net/azureml-blobstore-d49d62ee-fc90-4e2a-add1-39879de70ed8/embeddings/refined/split/)

## Related Issues

Closes #169 

## Additional Notes

We may have an issue with the split out embeddings as the format looks
different than the original tensor. We may need to update the notebook
to convert back to a tensor object prior to writing the split embedding
ljson files.... Will confirm.
## Checklist for Reviewers

Please review and complete the following checklist during the review
process:

- [ ] The code follows best practices and conventions.
- [ ] The changes implement the desired functionality or fix the
reported issue.
- [ ] The tests cover the new changes and pass successfully.
- [ ] Any potential edge cases or error scenarios have been considered.