Add Basic Configuration and Test for Text Viability, by Data Element, for TTC evaluation (#179)
## Description

This PR address the ability to do a variety of 'sanity'/'viability'
checks on the input text for TTC. This will ensure that before we
perform, the more expensive process of generating an embedding of a text
string and then evaluating within our model, we ensure that the text
isn't just garbage. To address this issue the following have been added:

- Config - I added to the existing config.py a list of proper data
elements as well as a dictionary of viability check rules per data
element. For now the only rule present is for word count of the text for
lab_order and lab_result.

- A new extraction.py - contains a set of functions specific to
extracting and prepping text strings from eCR messages for evaluation
within the TTC module. I created some sub functions that verify the data
field is proper and one for the specific test for word count evaluation,
if present for a data element within the config.py.

- Tests - I created tests for both the config.py and extraction.py

## Related Issues

Closes #156 

## Additional Notes

If there is a better way to handle this or more standard way, let me
know and I'm happy to adjust. I just tried to leverage some of the
patterns we already have within other sections of our code and wanted to
keep this as simple as possible.


## Checklist for Reviewers

Please review and complete the following checklist during the review
process:

- [ ] The code follows best practices and conventions.
- [ ] The changes implement the desired functionality or fix the
reported issue.
- [ ] The tests cover the new changes and pass successfully.
- [ ] Any potential edge cases or error scenarios have been considered.

---------

Co-authored-by: Josh Nygaard <josh.nygaard@gmail.com>