Extract Schematron Errors for TTC data fields of interest from APHL Schematron Output (#189)
## Description

Added a basic function to accept schematron output, from APHL, and parse
through to find 'errors' of interest to TTC. These are stored in a list,
by data field type, in our Configs. These error texts 'SHOULD' be
confirmed with APHL or this step will fail. If we can find a better way
to find errors of interest we should use it, but so far, based upon the
output provided by APHL, this seems to be the route.

Also, note that we will need to verify that the XPATHs provided will be
useful and work for us in the later steps. I'll attempt to test this in
the next stories. If there are issues, I will adjust this function to
adapt the XPATHs to what we need.

## Related Issues

Closes #171 

## Additional Notes

Note that we can modify how we are doing this, but we have something in
place for now.

## Checklist for Reviewers

Please review and complete the following checklist during the review
process:

- [ ] The code follows best practices and conventions.
- [ ] The changes implement the desired functionality or fix the
reported issue.
- [ ] The tests cover the new changes and pass successfully.
- [ ] Any potential edge cases or error scenarios have been considered.